feat(client): persist task status across session
