--- client/src/types/supabase.d.ts
@@ -1,4 +1,0 @@
-export interface SupabaseContext {
-	auth: SupabaseAuthClient;
-	user: User | null
-};
